Handle potential overflows in `parse_number`
This fixes an issue discovered by AFL.rs, where we eagerly unwrap the result of
`isize::from_str_radix` because we checked that the string contains only valid
digits. However, we did not check that the string of digits form a number within
the range of `isize`, and potentially try to parse a number larger than
`isize::MAX`. This resulted in an `Err` return value, and a panic when we unwrap
it.