Keep a proper stack of inner items during demangling
I was hoping to avoid this, but it doesn't seem like that is possible. This
replaces the `DemangleWithInner` trait with a stack of inner items for
demangling on the context. This required a ton of lifetime wrangling to
compile. We now statically distinguish between demangling as an outer
item (Demangle::demangle) vs demangling in an inner
position (DemangleAsInner::demangle_as_inner).

Additionally, this adds logging during demangling when the logging feature is
enabled.