Cache TemplateParameter to TemplateArg resolution
Remember which TemplateArg we mapped a TemplateParam to. We only have
the proper argument scope the first time we visit a TemplateArg, and
any substitution back references could be within a completely
different scope.  If we try to (re-)resolve a TemplateParam's concrete
TemplateArg outside the scope it was defined in, we'll get garbage
results.