Introduce the concept of "inner barriers".
Consider the mangled symbol `_ZSt4copyIPKcPcET0_`. This is correctly demangled to `char* std::copy<char const*, char*>()`. BecauseEncoding::Function places the function itself on the inner stack so that it can later be pulled off to decorate the full function name with arguments, if the template arguments happen to go looking on the inner stack they will find the function and pull in the function arguments as well. This leads to the nonsensical demangling `char* std::copy<char const*(), char*>`.

To prevent these sorts of problems, we introduce "inner barriers". Conceptually, these are nodes that a descendant looking back up the tree (via the inner stack) cannot see past. This is implemented through an RAII helper that sets aside the existing inner stack for an empty one at an inner barrier and replaces the previous inner stack when the helper goes out of scope. There's also an `inner_barrier` macro to make this transparent to `Demangle` implementations.

Similar to the template-args production, the encoding production can also contain complex elements within it that should not look back up the inner stack. There are probably others too ... a formal analysis of the grammar would help here.