Reorder parsing of type qualifiers to match the spec.
The spec says that the `qualified-type` production (within `type`) is
parsed second only to the `builtin-type` production. The `qualified-type`
production is ambiguous with later productions (in particular with
the "Ul" inside `class-enum-type`) so the order does matter. This
reorders things for "Ul" and brings along the other parts of the
`qualified-type` production.

To preserve the existing behavior for `function-type`, rather than attempt
to pass down a leading `CV-qualifiers` we simply skip them if the
tail remaining indicates a `function-type` production is imminent.